commit for Feature Branch. implementation of orderController & exception handling.

+++ src/main/java/org/springframework/web/Homework_38/service/OrderService.java
@@ -35,6 +35,12 @@ public class OrderService {
     }
 
     public void deleteOrder(int id) {
-        orderRepository.deleteById(id);
+        Order order;
+        Optional<Order> optionalOrder = orderRepository.findById(id);
+        if (optionalOrder.isPresent()) {
+            orderRepository.deleteById(id);
+        } else {
+            throw new RuntimeException();
+        }
     }
 }
